Move auth provider handler to config::auth
