+++ kube/src/config/auth.rs
@@ -6,11 +6,12 @@ use std::{
 
 use chrono::{DateTime, Utc};
 use http::header;
+use jsonpath_lib::select as jsonpath_select;
 use serde::{Deserialize, Serialize};
 use tokio::sync::Mutex;
 
-use super::{utils, AuthInfo, ExecConfig};
-use crate::error::ConfigError;
+use super::{utils, AuthInfo, AuthProviderConfig, ExecConfig};
+use crate::{error::ConfigError, oauth2, Result};
 
 #[derive(Debug, Clone)]
 #[allow(clippy::large_enum_variant)]
@@ -154,4 +155,128 @@ fn auth_exec(auth: &ExecConfig) -> Result<ExecCredential, ConfigError> {
     let creds = serde_json::from_slice(&out.stdout).map_err(ConfigError::AuthExecParse)?;
 
     Ok(creds)
+}
+
+pub(crate) async fn token_from_provider(provider: &AuthProviderConfig) -> Result<Option<String>> {
+    let mut token = None;
+    if let Some(access_token) = provider.config.get("access-token") {
+        token = Some(access_token.clone());
+        if utils::is_expired(&provider.config["expiry"]) {
+            // TODO This is GCP only. Check provider.name == "gcp"?
+            let token_res = oauth2::get_token().await?;
+            token = Some(token_res.access_token);
+        }
+    }
+
+    if let Some(id_token) = provider.config.get("id-token") {
+        token = Some(id_token.clone());
+    }
+
+    if token.is_none() {
+        if let Some(cmd) = provider.config.get("cmd-path") {
+            let params = provider.config.get("cmd-args").cloned().unwrap_or_default();
+
+            let output = Command::new(cmd)
+                .args(params.trim().split(' '))
+                .output()
+                .map_err(|e| ConfigError::AuthExec(format!("Executing {:} failed: {:?}", cmd, e)))?;
+
+            if !output.status.success() {
+                return Err(ConfigError::AuthExecRun {
+                    cmd: format! {"{} {}", cmd, params},
+                    status: output.status,
+                    out: output,
+                }
+                .into());
+            }
+
+            if let Some(field) = provider.config.get("token-key") {
+                let pure_path = field.trim_matches(|c| c == '"' || c == '{' || c == '}');
+                let json_output: serde_json::Value = serde_json::from_slice(&output.stdout)?;
+                match jsonpath_select(&json_output, &format!("${}", pure_path)) {
+                    Ok(v) if !v.is_empty() => {
+                        if let serde_json::Value::String(res) = v[0] {
+                            token = Some(res.clone());
+                        } else {
+                            return Err(ConfigError::AuthExec(format!(
+                                "Target value at {:} is not a string",
+                                pure_path
+                            ))
+                            .into());
+                        }
+                    }
+                    Err(e) => {
+                        return Err(
+                            ConfigError::AuthExec(format!("Could not extract JSON value: {:}", e)).into(),
+                        );
+                    }
+                    _ => {
+                        return Err(
+                            ConfigError::AuthExec(format!("Target value {:} not found", pure_path)).into(),
+                        );
+                    }
+                };
+            } else {
+                token = Some(
+                    std::str::from_utf8(&output.stdout)
+                        .map_err(|e| ConfigError::AuthExec(format!("Result is not a string {:?} ", e)))?
+                        .to_owned(),
+                );
+            }
+        } else {
+            return Err(ConfigError::AuthExec(format!(
+                "no token or command provided. Authoring mechanism {:} not supported",
+                provider.name
+            ))
+            .into());
+        }
+    }
+
+    Ok(token)
+}
+
+#[cfg(test)]
+mod test {
+    use crate::config::Kubeconfig;
+
+    use super::*;
+    #[tokio::test]
+    async fn exec_auth_command() -> Result<()> {
+        let test_file = "
+        apiVersion: v1
+        clusters:
+        - cluster:
+            certificate-authority-data: XXXXXXX
+            server: https://36.XXX.XXX.XX
+          name: generic-name
+        contexts:
+        - context:
+            cluster: generic-name
+            user: generic-name
+          name: generic-name
+        current-context: generic-name
+        kind: Config
+        preferences: {}
+        users:
+        - name: generic-name
+          user:
+            auth-provider:
+              config:
+                cmd-args: '{\"something\": \"else\", \"credential\" : {\"access_token\" : \"my_token\"} }'
+                cmd-path: echo
+                expiry-key: '{.credential.token_expiry}'
+                token-key: '{.credential.access_token}'
+              name: gcp
+        ";
+
+        let mut config: Kubeconfig = serde_yaml::from_str(test_file).map_err(ConfigError::ParseYaml)?;
+        let auth_info = &mut config.auth_infos[0].auth_info;
+        assert!(auth_info.token.is_none());
+        if let Some(provider) = &auth_info.auth_provider {
+            auth_info.token = token_from_provider(provider).await?;
+        }
+        assert_eq!(auth_info.token, Some("my_token".to_owned()));
+
+        Ok(())
+    }
 }
